Updated sequence number to int64

+++ message.go
@@ -9,13 +9,13 @@ import (
 //Message is an Azure Service Bus message
 type Message struct {
 	DeliveryCount          int
-	EnqueuedSequenceNumber int
+	EnqueuedSequenceNumber int64
 	EnqueuedTimeUtc        Time
 	LockToken              string
 	LockedUntilUtc         Time
 	MessageID              string `json:"MessageId"`
 	PartitionKey           string
-	SequenceNumber         int
+	SequenceNumber         int64
 	State                  string
 	TimeToLive             int
